Handle Non-Existent Contacts Table in Bootstrap Function in (exp… (#29)
* Fix: Handle Non-Existent Contacts Table in Bootstrap Function in (express.md)

This PR addresses an issue where the bootstrap function encounters an error if the contacts table doesn't exist in the database. The existing code attempts to query the contacts table and bootstrap the database if there are no rows, but it doesn't handle the case where the table itself is missing. 

Signed-off-by: Thierno Abdoul Rahimi Diallo <rahimi.diallo@224tech.com>

* Updating Express tutorial to handle cases where the database table doesn't exist yet

---------

Signed-off-by: Thierno Abdoul Rahimi Diallo <rahimi.diallo@224tech.com>
Co-authored-by: James Harrison <james@fl0.com>